requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==3.4.8
Django==4.2.13
django-bootstrap-v5==1.0.11
djangorestframework==3.15.2
idna==3.7
Naked==0.1.32
pybit==5.7.0
pycparser==2.22
pycryptodome==3.20.0
PyJWT==2.8.0
PyYAML==6.0.1
requests==2.32.3
shellescape==3.8.1
soupsieve==2.5
sqlparse==0.5.0
typing_extensions==4.12.2
urllib3==2.2.1
websocket-client==1.8.0
websockets==12.0
